+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Fixed
+- Add sample configuration file to deb and txz packages
+
 ## [2.0.0] - 2017-02-16
 ### Added
 - Subcommands to manage the backups (sync, get, list, remove, start, encrypt)